Refactor some things for cleanliness.
