+++ spoon-runner/src/main/java/com/squareup/spoon/html/HtmlDevice.java
@@ -1,5 +1,9 @@
-package com.squareup.spoon;
+package com.squareup.spoon.html;
 
+import com.squareup.spoon.DeviceDetails;
+import com.squareup.spoon.DeviceResult;
+import com.squareup.spoon.DeviceTest;
+import com.squareup.spoon.DeviceTestResult;
 import com.squareup.spoon.misc.StackTrace;
 import java.io.File;
 import java.util.ArrayList;
