+++ spoon-runner/src/main/java/com/squareup/spoon/html/HtmlIndex.java
@@ -1,5 +1,10 @@
-package com.squareup.spoon;
+package com.squareup.spoon.html;
 
+import com.squareup.spoon.DeviceDetails;
+import com.squareup.spoon.DeviceResult;
+import com.squareup.spoon.DeviceTest;
+import com.squareup.spoon.DeviceTestResult;
+import com.squareup.spoon.SpoonSummary;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedHashSet;
