+++ spoon-runner/src/main/java/com/squareup/spoon/html/HtmlUtils.java
@@ -1,5 +1,6 @@
-package com.squareup.spoon;
+package com.squareup.spoon.html;
 
+import com.squareup.spoon.DeviceTestResult;
 import com.squareup.spoon.misc.StackTrace;
 import java.io.File;
 import java.io.IOException;
