+++ spoon-runner/src/main/resources/page/device.html
