+++ spoon-runner/src/main/resources/static/icon-animated.png
