+++ spoon-runner/src/main/resources/static/icon-devices.png
