+++ spoon-runner/src/main/resources/static/icon-log.png
